+++ src/ApolloState.m
@@ -80,6 +80,7 @@ BOOL sPerAccountFavoritesEnabled = NO;
 BOOL sHideSubredditListDescriptions = NO;
 BOOL sHideMultiredditDescriptions = NO;
 BOOL sEnableFlairColors = NO;
+BOOL sBoldPostTitles = NO;
 BOOL sEnableInlineImages = NO;
 BOOL sEnableChatMedia = NO;   // effective default YES via registerDefaults (UDKeyEnableChatMedia)
 BOOL sEnableAISummaries = NO;